Restructuring the gem paths to be cleaner before pushing it to gemcutter.
