+++ init.rb
@@ -1,2 +1,1 @@
-ActionController::Base.send :include, RespondsToParent
-require 'parent_selector_assertion'
+require 'responds_to_parent'
