--- rails/init.rb
@@ -1,2 +1,0 @@
-ActionController::Base.send :include, RespondsToParent
-require 'parent_selector_assertion'
